% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWSB.R
\name{clipShape}
\alias{clipShape}
\title{Basic plot of WSB based on huc}
\usage{
clipShape(shapefile, mapRange)
}
\arguments{
\item{shapefile}{shapefile to clip}

\item{mapRange}{vector of map limits (min long, max long, min lat, max lat)}
}
\description{
Basic plot
}
\examples{
mapRange=c(-80,-74, 38, 46)
shape_hydropoly <- shape_hydropoly
clippedShape <- clipShape(shape_hydropoly, mapRange)
}

