% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWSB.R
\name{getFlowLines}
\alias{getFlowLines}
\title{Get shapefile flowlines}
\usage{
getFlowLines(Range, streamorder = 3, filePath = NA)
}
\arguments{
\item{streamorder}{integer stream order}

\item{filePath}{path to save shapefile. If NA, will go to temporary directory}

\item{mapRange}{vector of map limits (min long, max long, min lat, max lat)}
}
\value{
shapefile
}
\description{
Get shapefile flowlines
}
\examples{
\dontrun{
Range=c(-86.32679,-81.16322,39.61600,43.06262)
flowLines <- getFlowLines(Range, 5)
}
}

