% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.DEM.R
\name{import.DEM}
\alias{import.DEM}
\title{Reads in an ARCMAP ASCII grid file containing the DEM and prepares it for use in the head mapping.}
\usage{
import.DEM(grid, debug.level = 0)
}
\arguments{
\item{\code{grid}}{character string for the ARCMAP ASCII grid file name.}

\item{\code{debug.level}}{Control the user messages. A value >0 outputs progress.}
}
\value{
A gridded Spatial object will be returned with one field named "DEM".
}
\description{
\code{import.DEM} reading in a an ARCMAP ASCII grid file.
}
\details{
This function reads in an ARCMAP ASCII grid file containing the DEM
and prepares it for use in the mapping of groundwater level data. The preparations
include removing NA grid cells and converting it to a fully gridded Spatial object.
Also, the coordinates must be projected.
}
