% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.pointData.R
\name{import.pointData}
\alias{import.pointData}
\title{Reads in the point groundwater level observations file.}
\usage{
import.pointData(data, min.sepDist = 5, orderby.head = TRUE,
  debug.level = 0)
}
\arguments{
\item{\code{data}}{character string for the point data .csv filename. The file must contain columns with the following names "easting", "northing" and "head".}

\item{\code{min.sepDist}}{scalar value >0 defining the maximum distance between two observations that will result in a duplicate being detected. When a duplicate is detected then when
\code{orderby.head==TRUE}, the point with the greatest water level elevation will be kept ALso, the lowest elevation point will be kept. The default is 5m.

' @param \code{orderby.head} logical scalar defining if the highest elevation water level should be selected when a duplicate is found. The default is \code{TRUE}.}

\item{\code{debug.level}}{Control the user messages. A value >0 outputs progress.}
}
\value{
A point Spatial object  will be returned.
}
\description{
\code{import.pointData} reads in a .csv file of the point groundwater level data.
}
\details{
This function reads in .csv file of the point groundwater level data and
removing duplicates points. The coordinates must be projected.
}
