% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWSB.R
\name{plotWSB}
\alias{plotWSB}
\title{Basic plot of WSB based on huc}
\usage{
plotWSB(sites, col = "#A8A8A850")
}
\arguments{
\item{sites}{character vector of site ids}

\item{col}{for basin fill}
}
\description{
Basic plot
}
\examples{
sites <- c("01491000", "01573000", "01576000","01137500","06485500")
path <- system.file("extdata", package="hydroMap")
siteInfo <- readRDS(file.path(path,"siteInfo.rds"))
png("test.png")
plotWSB(sites)
points(siteInfo$dec_long_va, siteInfo$dec_lat_va, pch=20, col="red", cex=1)
box()
dev.off()
}

