% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWSB.R
\name{getBasin}
\alias{getBasin}
\title{Get shapefile basins}
\usage{
getBasin(sites)
}
\arguments{
\item{sites}{character id}
}
\value{
shapefile
}
\description{
Get shapefile basins
}
\examples{
sites <- c("01491000", "01573000", "01576000","01137500","06485500")
basinShapes <- getBasin(sites)
}

