% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWSB.R
\name{getFlowLines}
\alias{getFlowLines}
\title{Get shapefile flowlines}
\usage{
getFlowLines(Range, streamOrder = 3)
}
\arguments{
\item{mapRange}{vector of map limits (min long, max long, min lat, max lat)}

\item{streamorder}{integer stream order}
}
\value{
shapefile
}
\description{
Get shapefile flowlines
}
\examples{
Range=c(-86.32679,-81.16322,39.61600,43.06262)
flowLines <- getFlowLines(Range, 5)
}

