% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWSB.R
\name{plotWSB}
\alias{plotWSB}
\title{Basic plot of WSB based on huc}
\usage{
plotWSB(sites, col = "#A8A8A850", mapRange = NA, flowDensity = 3)
}
\arguments{
\item{sites}{character vector of site ids}

\item{col}{for basin fill}

\item{mapRange}{vector of map limits (min long, max long, min lat, max lat)}
}
\description{
Basic plot
}
\examples{

library(dataRetrieval)
Range=c(-86.32679,-81.16322,39.61600,43.06262)
sites=c("04189000","04197100","04198000","04185000","04199500","04176500","04193500")
basins=getBasin(sites)
siteInfo <- readNWISsite(sites)

png("test.png",width=11,height=8,units="in",res=600,pointsize=4)
plotWSB(sites, mapRange=Range)
points(siteInfo$dec_long_va, siteInfo$dec_lat_va, pch=20, col="red", cex=3)
box()
dev.off()
}

