% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWSB.R
\name{plotWSB}
\alias{plotWSB}
\title{Basic plot of WSB based on huc}
\usage{
plotWSB(sites)
}
\arguments{
\item{siteInfo}{data frame returned from dataRetrieval}

\item{rangeMap}{numeric vector of: min longitude, max longitude, min latitude, max latitude}
}
\value{
list of politicalBoundary, rivers, and lakes shapefiles, and a vector of lat/lon limits
}
\description{
Basic plot
}
\examples{
sites <- c("01491000", "01573000", "01576000","01137500","06485500")
png("test.png")
plotWSB(sites)
points(siteInfo$dec_long_va, siteInfo$dec_lat_va, pch=20, col="red", cex=3)
box()
dev.off()
}

