% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWSB.R
\name{plotWSB}
\alias{plotWSB}
\title{Basic plot of WSB based on huc}
\usage{
plotWSB(sites, col = "#A8A8A850", mapRange = NA, streamorder = 3,
  filePath = NA)
}
\arguments{
\item{sites}{character vector of site ids}

\item{col}{for basin fill}

\item{mapRange}{vector of map limits (min long, max long, min lat, max lat)}

\item{streamorder}{integer}

\item{filePath}{path to save shapefiles. If NA, will go to temporary directory}
}
\description{
Basic plot
}
\examples{
\dontrun{
library(dataRetrieval)
Range=c(-86.32679,-81.16322,39.61600,43.06262)
sites=c("04189000","04197100","04198000","04185000","04199500","04176500","04193500")
siteInfo <- readNWISsite(sites)

png("test.png",width=11,height=8,units="in",res=600,pointsize=4)
plotWSB(sites, mapRange=Range)
points(siteInfo$dec_long_va, siteInfo$dec_lat_va, pch=20, col="red", cex=3)
box()
dev.off()


site <- '08076000'
siteInfo <- readNWISsite(site)
png("test.png",width=11,height=8,units="in",res=600,pointsize=4)
plotWSB(site)
points(siteInfo$dec_long_va, siteInfo$dec_lat_va, pch=20, col="red", cex=3)
box()
dev.off()
}
}

